✨🛠  Enable the API and WebSockets within vue-cli, and prod build
