+++ package.json
@@ -22,6 +22,7 @@
   "homepage": "https://github.com/omgmog/sound-dropper-server#readme",
   "dependencies": {
     "bulma": "^0.9.2",
+    "connect-history-api-fallback": "^1.6.0",
     "express": "^4.17.1",
     "get-audio-duration": "^2.0.3",
     "howler": "^2.2.1",
