+++ vue.config.js
@@ -1,0 +1,7 @@
+const configureAPI = require('./server/api')
+
+module.exports = {
+  devServer: {
+    before: configureAPI
+  }
+}
